docs(README.md): update instructions for running commands and clarify the purpose of the scripts
The instructions for running commands have been updated to reflect the
location of the scripts in the `src/deploy` folder. The commands `yarn
synth` and `yarn deploy` are used to run these scripts. The purpose of
the `deploy` command is clarified by mentioning that it also runs
`synth`, but it is easier to verify the `synth` command separately.