docs(arangodb_database.md): add documentation for Arangodb database configuration options
The commit adds a new file `arangodb_database.md` to the documentation
folder. This file provides documentation for various command line
options that can be used to configure the Arangodb database
installation. The documentation includes descriptions and default values
for each option, as well as whether the option is required or not.