fix(dictycr.ts): change 'database' variable name to 'logLevel' for better semantics
The 'database' variable in the 'BackendDeploymentResource' and
'containerProperties' types has been renamed to 'logLevel' to improve
clarity and reflect its actual purpose. The corresponding changes have
been made in the 'BackendDeployment' class as well, including the method
names and arguments. The 'database' argument in the '#commandArgs'
method has been replaced with 'logLevel' to accurately represent the
command arguments.