feat(install_backend_app.ts): add script to install backend app with configurable options
The `install_backend_app.ts` script is added to the project. This script
allows for the installation of the backend app with various configurable
options. These options include the Kubernetes config file, service
account credentials file for GCS backend, GCS bucket name for storing
Terraform remote state, whether to use the remote GCS backend, name of
the installation, Kubernetes namespace to install, dictycr secret
containing the backend credentials, image and tag for the backup job,
log level of the app, and the port number for the service. The script
creates a backend deployment and service using the provided options.