fix(install_nats.ts): update app attribute in NatsBackendService to match the value of name attribute passed in the helm chart
The describe messages for the 'nm' and 'ns' options have been updated to
provide clearer explanations of their purposes. The name attribute in
the HelmChartStack has been updated to match the app attributes passed
in the service. The app attribute in NatsBackendService has been updated
to match the value of the name attribute passed in the helm chart.