feat(install_nats.ts): add functionality to list services in a Kubernetes namespace
The `listServices` function is added to retrieve a list of services in a
Kubernetes namespace using the provided kubeconfig and namespace. This
functionality can be useful for various deployment and management tasks.
Additionally, a new option `-l` or `--level` is added to specify the
logging level for the installation process. This allows for more
granular control over the amount of logging output during the
installation of NATS.