feat(deploy): add script to create dictycr configuration
The `create_dictycr_config.ts` script is added to the `src/deploy`
directory. This script allows for the creation of a configuration for
the `dictycr` application. It uses the `yargs` library to parse command
line arguments and create a configuration map stack using the
`ConfigMapStack` class from the `../construct/dictycr` module. The
script accepts various options such as the name of the installation, the
Kubernetes namespace, the Kubernetes config file, service account
credentials file for GCS backend, GCS bucket name for storing Terraform
remote state, whether to use the remote GCS backend, and API endpoints
for organism and publication. The script creates the configuration map
stack and synthesizes the app.