feat(install_graphql_server.ts): add script to install and configure a GraphQL server deployment
The `install_graphql_server.ts` script is added to the project. This
script allows for the installation and configuration of a GraphQL server
deployment in a Kubernetes cluster. It takes in various command line
options such as the Kubernetes config file, service account credentials
file, GCS bucket name for Terraform remote state, deployment name,
namespace, image, tag, log level, port number, and allowed http origins.
The script uses the `GraphqlBackendDeploymentStack` and `BackendService`
constructs to create the necessary resources for the deployment and
service.