feat(postgres.ts): add support for dynamic user and database creation in PostgresStack class
The PostgresStack class now accepts a user and an array of databases as
options. This allows for dynamic creation of users and databases in the
PostgresStack. The #users method has been updated to use the provided
user and databases values when creating the user object. This
enhancement improves the flexibility and configurability of the
PostgresStack class.