feat(install_pg_schema.ts): add script to install PostgreSQL schema using Terraform CDKTF
The `install_pg_schema.ts` script is added to the
`src/deploy/management` directory. This script allows for the
installation of a PostgreSQL schema using Terraform CDKTF. It takes in
various command line options such as the name of the installation, the
Kubernetes namespace to install, the Kubernetes config file, the GCS
bucket name for storing the Terraform remote state, and the PostgreSQL
secret and database information. It also allows for specifying the image
and tag to use for installing the schema. The script creates a new
instance of the `PgschemLoadingStack` construct from the `schema_job.ts`
file and synthesizes the CDK app.