feat(install_event_messenger.ts): add script to install event messenger backend
The `install_event_messenger.ts` script is added to the project. This
script allows for the installation of the event messenger backend. It
takes in various command line options such as the Kubernetes config
file, service account credentials file, GCS bucket name, whether to use
a remote GCS backend, Kubernetes namespace, name of the installation,
image and tag to use for the backend, log level, secret name, config map
name, and NATS subject name. The script creates an instance of the
`EmailBackendDeploymentStack` and deploys the event messenger backend
using the provided options.