feat(dictycr.ts): add support for graphlEndpoint in ConfigMapStack constructor
The ConfigMapStack constructor now accepts a new property called
graphlEndpoint in the options parameter. This allows the ConfigMapStack
to configure the "eventmessenger.endpoint.publication" key in the data
section of the ConfigMap with the provided graphlEndpoint value. This
change enables the application to dynamically set the graphlEndpoint
value based on the environment or configuration.