feat(logto.ts): add support for an init container to run a CLI command during deployment
The LogtoBackendDeploymentStack class now includes support for an init
container that runs a CLI command during deployment. This init container
is responsible for adding an official connector using the npm run cli
connector add command. The command is executed inside the container
using /bin/sh and the necessary volume mount is configured to access the
required files.