fix(logto.ts): change diskSize type from string to number for better type safety
The diskSize property in the LogtoPersistentVolumeClaimStackProperties
type has been changed from a string to a number to improve type safety.
Additionally, there was a typo in the accessModes value, which has been
corrected to "ReadWriteOnce". Lastly, the diskSize value in the
requests.storage property was missing the "Gi" suffix, which has been
added to ensure correct formatting.