feat(install_logto.ts): add support for specifying the name of the logto database via the --db flag
The --db flag has been added to the command line options for the
install_logto middleware. This allows users to specify the name of the
logto database when deploying the LogtoBackendDeploymentStack. This
provides flexibility and customization options for users who want to use
a different name for their logto database.