refactor(logto.ts): remove unnecessary alteration deploy command and update DB_URL environment variable
The alteration deploy command was removed from the args array as it is
no longer needed. The DB_URL environment variable was updated to include
the "?sslmode=no-verify" query parameter for better security.
Additionally, a new environment variable PGPASSWORD was added to store
the decoded password from the secret data.