feat(logto_types.ts): add new type logtoIngressArgvProperties and export it
The commit adds a new type `logtoIngressArgvProperties` to the
`logto_types.ts` file. This type extends the existing
`providerArgvProperties` and `basicArgvProperties` types and adds
additional properties `sr`, `sc`, `is`, and `hs`. The commit also
updates the export statement to include the newly added type. This
change is done to provide support for a new set of command line
arguments related to logging.