feat(logto.ts): introduce LogtoIngressStack class to create an Ingress resource for the Logto application
The LogtoIngressStack class is added to the logto.ts file to provide
support for creating an Ingress resource in the Logto application. This
allows the application to be exposed externally and handle incoming
traffic. The Ingress resource is created using the Kubernetes
networking.k8s.io/v1 API version and is configured with the necessary
metadata, annotations, and rules. The LogtoIngressStack class is
exported to be used in other parts of the application.