feat(helm.ts): add optional location property to HelmChartProperties
The addition of the optional location property to HelmChartProperties
allows for more flexibility in specifying the location of the helm
chart. The refactoring of the Release instantiation now supports the use
of this new location property. If a location is provided, it is used as
the chart's location. If not, the repository and name properties are
used as before. This change allows for more diverse helm chart sources.