feat(logto.ts): extend startup script for container
The startup script for the container has been extended to include
additional commands for database alteration deployment and connector
linking. This ensures that the necessary setup steps are performed each
time the container starts, improving the reliability of the application.
The readOnly restriction from volumeMount has been removed to allow
write operations. This change provides the application with the
necessary permissions to modify files in the
'/etc/logto/packages/core/connectors' directory, which can be essential
for certain operations.